validation: remove TopGit management
This is the dedicated job of the original action.

Here we now have a new Github Action to do the validation. The goal is
to do that in parallel jobs to accelerate the validation on limited
resources machines.

A few configurations can be given. 3 of them are interesting for a
matrix mode:

  - defconfig: x86_64 or i386
  - ipv6: with_ipv6 or without_ipv6
  - mptcp: with_mptcp or without_mptcp

We can also pass 'base' to indicate if we want to validate each commit
of the export branch or only commits on top of it.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>