compile: no skip with each_commit:0
If we only want to validate the last commit, we should not skip a
compilation depending on the diff with the parent commit.

If we use 'each_commit: 0', it likely means we don't know the base
(COMMIT_BOTTOM). We just want to validate the last commit and check that
everything compiles properly.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>