gh action: push new images to docker hub
Autobuilds is no longer available for us:

  https://www.docker.com/blog/changes-to-docker-hub-autobuilds/

An alternative is then to add a Github Action to build the docker image
when a modification is done in this repo and push it to hub.docker.com.

The content of the YAML file is highly inspired by the example:

  https://github.com/marketplace/actions/build-and-push-docker-images

I don't think we need the whole machinery (e.g. QEmu) but easier to
follow the recommended steps.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>